+++ include/map.h
@@ -1,7 +1,7 @@
 #ifndef MAP_H
 #define MAP_H
 
-#define MAP_POS(x, y, z) ((int)((x) + ((y) * m_map->width) + ((z) * m_map->width * m_map->depth)))
+#define MAP_POS(x, y, z) ((x) + ((y) * m_map->width) + ((z) * m_map->width * m_map->depth))
 
 typedef struct {
 	u16 width;
